Fixed an issue with the J MIPS
the (2 || 3) was not working. It wasn't erroring since technically that evaluates to a integer, but it never evaluates to 2 or 3 and therefore would end up going through the default (the I case) and when it went through the I case it would go to the default of that which is the Unkown OP Code error.